Used numpy to have more control over datatypes.
Rewrote some of the code such that numpy can be used to control the datatypes.
I want the shingle function to return the shingles hashed as 4byte integers,
because it saves space and the algorithm still works with it, because a few
collision do not matter. Since python does not allows you to use datatypes
natively I used the types numpy provided.
Because, I used numpy for this, I also decided to use the numpy for the other
datastructures.